backend/requirements.txt
Django==1.11.20
djangorestframework==3.9.4
django-cors-headers==3.0.2
